+++ src/main/scala/org/podval/tools/besom/User.scala
@@ -5,4 +5,4 @@ final class User(val name: String, val domain: String):
   def resourceName: String = s"user:$email"
   
 object User:
-  def gmail(name: String): User = User(name, "gmail.com")  
+  def gmail(name: String): User = User(name, "gmail.com")